ceia-proyecto-final/modulo-aplicaciones-web/prediction-app/backend/fastapi-backend/pyproject.toml
[project]
name = "fastapi-backend"
version = "0.1.0"
description = ""
authors = [{ name = "brunomaso1", email = "brunomaso1@hotmail.com" }]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "fastapi[standard]",
    "python-multipart",
    "python-dotenv",
    "loguru",
    "geopandas (>=1.1.1,<2.0.0)",
    "numpy (>=2.2.6,<3.0.0)",
    "opencv-python (>=4.11.0.86,<5.0.0.0)",
    "torch (>=2.7.1,<3.0.0)",
    "torchvision (>=0.22.1,<0.23.0)",
    "ultralytics (>=8.3.169,<9.0.0)",
    "fastkml (>=1.1.0,<2.0.0)",
    "geojson (>=3.2.0,<4.0.0)",
    "modulo-apps @ file:///E:/Documentos/Git%20Repositories/uba-ceia-proy-final/ceia-proyecto-final/modulo-apps",
    "modulo-ia @ file:///E:/Documentos/Git%20Repositories/uba-ceia-proy-final/ceia-proyecto-final/modulo-ia",
    "fastapi-keycloak-middleware (>=1.3.0,<2.0.0)",
    "fastapi-keycloak (>=1.1.0,<2.0.0)",
]

[tool.poetry]
packages = [{ include = "fastapi_backend", from = "src" }]

[[tool.poetry.source]]
name = "pytorch_gpu"
url = "https://download.pytorch.org/whl/cu128"
priority = "explicit"

[tool.poetry.dependencies]
torch = { source = "pytorch_gpu" }
torchvision = { source = "pytorch_gpu" }
modulo-apps = {develop = true}
modulo-ia = {develop = true}

[tool.poetry.group.dev.dependencies]
taskipy = "^1.14.1"

[tool.taskipy.tasks]
dev = "fastapi dev ./src/fastapi_backend/main.py"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
